requirements.txt
# Pin NumPy to a compatible version first (must be <2.0)
numpy<2.0

# Core dependencies
flask
python-dotenv
pydub
openai-whisper
pywebview

# ML/Audio processing
torch
pyannote.audio
scipy

# API integrations
requests
openai

# For bundling ffmpeg binary
imageio-ffmpeg